[servlet] Move mappers to servlet package
